refactor: rover initial position abscissa is now an unsigned integer

+++ src/main/kotlin/io/github/xstefanox/marsrover/MarsRover.kt
@@ -20,12 +20,12 @@ import java.lang.invoke.MethodHandles
 
 // TODO x type should be UInt
 // TODO y type should be UInt
-class MarsRover(x: Int = 0, y: Int = 0, direction: Direction = North, planet: Planet = Planet(1u, 1u)) {
+class MarsRover(x: UInt = 0u, y: Int = 0, direction: Direction = North, planet: Planet = Planet(1u, 1u)) {
 
     private val log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass())
 
     // TODO remove the integer conversion
-    var position: Position = Position(x.toUInt(), y.toUInt())
+    var position: Position = Position(x, y.toUInt())
         private set(value) {
             Position(value.x, value.y)
             field = value
@@ -110,7 +110,7 @@ class MarsRover(x: Int = 0, y: Int = 0, direction: Direction = North, planet: Pl
             }
 
             // TODO remove the integer conversion
-            return MarsRover(x.toInt(), y.toInt()).right()
+            return MarsRover(x, y.toInt()).right()
         }
 
         sealed class CreationFailure {
